+++ src/interfaces/AuthChecker.ts
@@ -1,8 +1,17 @@
+import { ClassType } from "./ClassType";
 import { ResolverData } from "./ResolverData";
 
-export type AuthChecker<ContextType = {}, RoleType = string> = (
-  resolverData: ResolverData<ContextType>,
-  roles: RoleType[],
+export type AuthCheckerFn<TContextType = {}, TRoleType = string> = (
+  resolverData: ResolverData<TContextType>,
+  roles: TRoleType[],
 ) => boolean | Promise<boolean>;
+
+export type AuthCheckerInterface<TContextType = {}, TRoleType = string> = {
+  check(resolverData: ResolverData<TContextType>, roles: TRoleType[]): boolean | Promise<boolean>;
+};
+
+export type AuthChecker<TContextType = {}, TRoleType = string> =
+  | AuthCheckerFn<TContextType, TRoleType>
+  | ClassType<AuthCheckerInterface<TContextType, TRoleType>>;
 
 export type AuthMode = "error" | "null";